// src/actions/types.ts
export enum actionTypes {
    authenticate,
    logout,
    authError,
    authLoading,

    showAlert,
    hideAlert,
    setNotifications,
    setFriends,
    setPendingInvitations,
    setOnlineUsers,
    setUserNotifcations,
    setGroupChatList,
    resetFriends,

    setChatType,
    setChosenChatDetails,
    setMessages,
    setGroupNotifications,
    addNewMessage,
    resetChat,
    setChosenGroupChatDetails,

    setTyping,
    setInitialTypingStatus,

    setLocalStream,
    setRemoteStream,
    setOtherUserId,
    setAudioOnly,
    setScreenSharingStream,
    setScreenSharing,
    setCallRequest,
    setCallStatus,
    resetVideoChatState,

    openRoom,
    setRoomDetails,
    setActiveRooms,
    setLocalStreamRoom,
    setRemoteStreams,
    setAudioOnlyRoom,
    setScreenSharingStreamRoom,
    setIsUserJoinedWithAudioOnly,
}

export interface CurrentUser {
    _id: string;
    email: string;
    token: string;
    username: string;
}
interface AuthSuccessAction {
    type: actionTypes.authenticate;
    payload: {
        _id: string;
        email: string;
        token: string;
        username: string;
    };
}

interface AuthErrorAction {
    type: actionTypes.authError;
    payload: string;
}
export interface OnlineNotifaction {
    department: string;
    totalNotifications: number;
}
interface LogoutAction {
    type: actionTypes.logout;
}
interface setNotifications {
    type: actionTypes.setNotifications;
    payload: Array<OnlineNotifaction>;
}

interface AuthLoadingAction {
    type: actionTypes.authLoading;
    payload: boolean;
}

interface ShowAlertAction {
    type: actionTypes.showAlert;
    payload: string;
}

interface HideAlertAction {
    type: actionTypes.hideAlert;
}

export interface PendingInvitation {
    _id: string;
    senderId: {
        username: string;
        email: string;
        _id: string;
    };
}

export interface Friend {
    id: string;
    name: string;
    username: string;
    email: string;
    department: string;
    section: string;
    designation: string;
}

export interface OnlineUser {
    userId: string;
    socketId: string;
}

export interface GroupChatDetails {
    groupId: string;
    groupName: string;
    participants: Array<{
        _id: string;
        username: string;
        email: string;
    }>;
    admin: {
        _id: string;
        username: string;
        email: string;
    };
}
export interface UserGetNotification {
    userId: string,
    messages:Array<[]>;
}

interface SetPendingInvitationAction {
    type: actionTypes.setPendingInvitations;
    payload: Array<PendingInvitation>;
}

interface SetFriends {
    type: actionTypes.setFriends;
    payload: Array<Friend>;
}

interface SetOnlineUsers {
    type: actionTypes.setOnlineUsers;
    payload: Array<OnlineUser>;
}
interface setUserNotifcations {
    type: actionTypes.setUserNotifcations;
    payload: Array<UserGetNotification>;
}

export interface SetGroupChatList {
    type: actionTypes.setGroupChatList;
    payload: Array<GroupChatDetails>;
}
// ...
export interface ResetFriends {
    type: actionTypes.resetFriends;
}
// ...
interface setOtherUserId {
    type: actionTypes.setOtherUserId;
    payload: {
        otherUserId: string | null;
    };
}

interface setScreenSharingStream {
    type: actionTypes.setScreenSharingStream;
    payload: {
        stream: MediaStream | null;
        isScreenSharing: boolean;
    };
}
// ...
export type FriendsActions =
    | SetPendingInvitationAction
    | SetFriends
    | SetOnlineUsers
    | setUserNotifcations
    | SetGroupChatList
    | ResetFriends
    | setNotifications;